Create schema.js, move legacy db code
