+++ package.json
@@ -37,11 +37,13 @@
     "mongodb": "^3.6.0",
     "mongoose": "^5.9.27",
     "nodemon": "^2.0.4",
+    "performance-now": "^2.1.0",
+    "pg": "^8.3.3",
     "prop-types": "^15.7.2",
     "react": "^16.13.1",
     "react-dom": "^16.13.1",
-    "styled-components": "^5.1.1",
-    "performance-now": "^2.1.0"
+    "sequelize": "^6.3.4",
+    "styled-components": "^5.1.1"
   },
   "devDependencies": {
     "@babel/core": "^7.11.1",
